Enabled SQL Server usage in prototype.
* Added to the main solution: A HACKED VERSION
DataSetInDataGrid.Silverlight project from the implementation
investigation.  The HACK was to convert the original project from a
Silverlight APPLICATION to a Silverlight LIBRARY, which turned out okay,
it was accomplished by modding the csproj XML file.
* Created a mockup DataSetGrid control in the prototype Screens
Silverlight Library that used the DataSetInDataGrid Library.
* Added the DataSetGrid control to the Home page for testing.
It works.
Now, on to more exciting things... Oi Vey.